change uint64_t to uintptr_t where it's supposed to hold a pointer value

+++ deps/libparsemaps/libparsemaps.h
@@ -4,7 +4,7 @@
 #include <stdint.h>
 
 struct mapping_info {
-	uint64_t start, end;
+	uintptr_t start, end;
 	int perms;
 	int private;
 	int offs;
